Se agregaron nueos componentes ala structurade stylus y los media quieries para el diseño de la webapp

+++ app/static/js/nota/NotaController.js
@@ -43,8 +43,10 @@
 			NotaService.updateNote($scope.nota,function(nota){});
 		};
 
-		$scope.delete = function(id){
-			NotaService.delete(id);
+		$scope.delete = function(id) {
+			if(confirm("¿Esta Seguro que desea eliminar esta nota?")){
+				NotaService.delete(id);
+			}
 		};
 
 		$scope.clear_model= function(){
